+++ src/main/java/com/sip/ams/repositories/ProviderRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sip.ams.entities.Provider;
 
 @Repository
-public interface ProviderRepository extends CrudRepository<Provider,Long> {
+public interface ProviderRepository extends CrudRepository<Provider, Integer> {
 
 }
